Reduce the number of files
Implement ICE inside _partial_dependence.py. Move ICE tests to test_partial_dependence.py and test_plot_partial_dependence.py.
